Extract alias method order logic
